Refactor code. Pull service call to listing.
